Add code examples
While the sketches do provide some code examples, they are pretty messy,
since they're more of a braindump than anything else, and they also
don't have the most descriptive names on the earth
(rubberband-sketch_n); all that makes it pretty hard to find a specific
example of some feature of the language being used. Because of this, I
decided to create the codeexamples directory, which will be a more
organized place to check for examples, since the files actually tell
what they are. Also, if anything in the language syntax changes, those
files will be updated accordingly -- unlike the stuff in sketches, which
mixes very, very old sketches and very, very experimental features.

Hope that helps.