Make blocks answer copies instead of self
Previously, when a block without a defined context received a message
it would set the message as its context and answer self. However,
to allow blocks to be used as templates for new objects, its better
to answer a copy and make the contexts of each copy independent from
each other.