Use follows_interface to determine if msg is valid
Before this commit, to determine if, for instance, an array
received a number (and not a, say, block), the array_send_msg
function would access the __value.type field. Doing so rendered
the metainfo system useless, since even if the object responded
to the same interface as a number, it would cause a syntax error.
Now, the interface info is used, even for native objects.