Fix not all paths returning a value
This issue didn't affect the unoptimized build, since the default behavior
is to default to zero when a return expression isn't reached at the end
of a function, but this doesn't always apply in the optmized build.

It's also interesting that this problem went unnoticed because neither
clang nor gcc issued any warnings about this. Compiler bug?